Make SavedSearchesController have an AS::Concern
This will enable removal of the monkey patching that BL::AdvancedSearch
is doing, which is causing the bug
https://github.com/projectblacklight/blacklight_advanced_search/issues/49
